feat(proxy): Added basic symstore proxy mode (#31)
This lets one configure a list of sources in the config file which is used
as a default for the symbolicate request. It also adds a new proxy mode which
is on by default which will serve up symbols downloaded from one of the source
servers. It will always decompress the symbols which lets one use this with
visual studio and other debuggers as a source even if the source server format
is not necessarily the correct format.