feat(config): Default cache_dir to /data on Docker (#110)
On our Docker images, we already define `/data` as a persistent volume, even more, it is quite tricky to mount a new cache directory as a volume with the right symbolicator user so make that the default on Docker. This also allows us to run symbolicator with the default config in getsentry/onpremise#220.