fix: Unified source lookups for Darwin Minidumps (#208)
Fixes lookups for images referenced from macOS Minidumps from unified
sources.

For MachO images, the Breakpad processor does not emit a code
identifier. However, symbolic defines the code identifier as the
contents of LC_UUID, which is equivalent to the debug identifier. This
is required by the unified source, since it always keys files by the
code identifier.

The byproduct of this is that we now backfill code identifiers into
debug images of macOS minidumps. This doesn't yield more information,
but it creates parity with the other platforms. Note that we do not yet
cross-convert debug ids and code ids for MachO files.