Reduce the threadpool sizes (#216)
Reduce the threadpool sizes

The most important here is the CPU-intensive threadpool which is no
longer the same size all the CPUs available.  The IO-bound threadpools
are still overcommited in that there are two threads per CPU core.

The num_cpus crate was already an indirect dependency, now it is a
direct dependency under the same version number as before.

Lastly add a cude measure of the healthchecks in the form of a metrics
counter.  This is not ideal as it's the server's side point of view
but will give some indication.  I'm be tempted to add the timestamp as
a field but not sure this makes much sense in Datadog's model.