Enable metrics in the spawned subprocesses
This configures statsd in the spawned subprocesses as well as in the
main server process.  This means subprocesses will be sending metrics
to the statsd agent as well.

As a side-effect the metrics will also be initialised when running the
cleanup jobs, which was not the case yet.  There are currently no
metrics logged in the cleanup job code-path, so practically this will
make no difference.