Create and inject the download service
This actually creates the download service and passes it through to
the object cache who needs this.  It would have been nice to pass a
more generic service registry instead of the download service
explicitly, but this registry is ServiceState and we can not pass it
to things before it is constructed.