fix: Do not report false-positives in demangling (#230)
This fixes a particular case where we reported that the C++ symbol
`_Task_impl` cannot be demangled.

This happens because `_T` is detected as `Swift` symbol. The original
debug file declared it as `C++`, meaning its an unmangled name in C++
code. In this case, detection resolves a false-positive which should be
ignored.