feat(minidump): Save minidumps which crash during processing (#233)
feat(minidump): Save minidumps which crash during processing

This adds a feature which saves minidumps when the process processing
them crashes.  It also improves the metrics reporting these failures
to contain more information about the failure.

The crashed minidumps are stored in the cache directory and have a
default cache timeout of 24h.