Allow better defauls and unlimited caches
This primarily splits the CacheConfig struct into 3 separate ones so
each can implement Default.  This means serde can correctly
deserialise all these with defaults and still represent None.

The original CacheConfig is truned into a struct and it's fields are
now accessors which is the interface which Cache actually uses.