Initial sketch of download service (#228)
ref: Move download code to a new download service

This introduces and explores a new services architecture for the downloading of
DIF files.  The downloader runs on it's own thread and no work is done on the
executor of the caller.  Because the current http client is not Send the executor
of the downloader itself had to be single-threaded which is introduced in the
RemoteThread struct.