Implement timeout and metrics on the RemoteThread
All users spawning onto the RemoteThread (or ThreadPool) should be
forced to do these things, so may as well force this into the API.
The previous macro also only works for futures01 and we need futures03
support.  Finally this does not have the bug of the macro which
double-counts tasks which time out.