fix: Create cache directories at startup (#238)
fix: Create cache directories at startup

The diagnostics cache directory was not created.  It seems all the
cache directories are created on the fly at runtime.  This changes
this to create all the cache directories at startup, ensuring the
application fails to start when it can not create one of the caches.
It also makes sure to create it at the point of use as well just in
case the directory had been removed.