Actually keep modules without a debug id
This leaves the modules with incomplete info in the modules list
returned to the caller.  Restores the stripping from modules when
retrieving the list so that we do not have to keep the, arguably
brittle, code around that then also tries to avoid looking these up in
case some symbol server returns data for them.