fix: Exclude irrelevant modules from minidumps (#255)
Follow-up to #247. Minidumps can specify shared memory regions in the
module list. These are undesired for symbolication purposes. The only
exception is when a stack frame references an address from such a
module.

This PR introduces `CodeModulesBuilder` which allows to track valid
modules and references into them. In the end, only modules with valid
identifiers or with references are kept.