Do not allow skipping tests with credentials on CI
When running on CI we really shouldn't accidentally skip tests, so we
add a new --ci option which does not allow skipping tests when this
option is present.

The makefile needs to work both for users as for the CI so add in a
backdoor to get commandline arguments into there.