feat(cache): Write temporary files in a dedicated directory (#265)
This changes where temporary files are created to be in a dedicated
directory instead of the destination.  The final destinations are pruned
using the `symbolicator cleanup` command and this does not expect
files to be deleted by anything but itself, thus erroring if temporary
files have been moved to their final locations.

Since the new location is still within the cache directory moves from
there to their final cache location are atomic as users are expected
to keep the entire cache directory on one filesystem.