fix(objects): Make sure returned files can be used for purpose
When the objects actor downloads files it ranks them and chooses the
highest ranking file.  However before returning we must make sure the
file can actually be used for the requested purpose.

In particular this avoids asking for CFI but only getting a (64-bit)
PDB but no PE file.  This file does not actually provide the requested
info, which then confuses the CfiActor because it tries to use it as a
CFI file, even though the file does not have CFI.