test: Move conftest.py to repo toplevel (#284)
Because this does things globally to the pytest process it needs to be
loaded before any test collection.  Thus it should be in the topmost
conftest.py, in this case the toplevel of the repo.

This currently works on CI because it invokes pytest with the path to
the integration tests, triggering the peculiarities of pytest plugin
loading which will now load a conftest in this directory before
collecting tests.  However when you interactively run pytest you might
not want to bother passing this argument all the time, let alone
realise this is the problem why pytest mysteriously fails.