test: Be more liberal when skipping credentials (#288)
When an external person makes a PR they do not get access to the
credentials in the environment variables and the tests will fail.  So
we now check we're actually running as getsentry to still allow
skipping for external contributors.

This also removes the --ci flag, as we are now depending on Github
Actions environment variables we might as well depend on them to
detect running in CI in general.