ci: Run tests with rust nightly and beta channels on a schedule
This runs only the unittests once a day using the nightly and the beta
channel.  This should not block PRs but will give us earlier warnings
when we need to adjust things.

No caching since the compiler will be different every day anyway and
it is not blocking PRs anyway.