Adopt rc_buffer clippy lint
This follows the clippy lint to not put a Vec (or String) in an Arc.
Doing this means we carry around the likely extra unused storage space
for the Vec.  Instead we can use Arc::from to create an Arc::<[T]>
directly which avoids the storage overheads.

Since this lint is no longer on by default in the next release it is
also manually enabled.