ref: Updates for Rust 1.48 (#297)
This requires a new insta because one of its dependencies needed to be
fixed.  In turn this changes the names of the snapshots being
generated.  The snapshots themselves do not change.

Intra-doc linking is now stable, no need to run rustdoc on nightly
anymore.

This also follows the clippy lint to not put a Vec (or String) in an Arc.
Doing this means we carry around the likely extra unused storage space
for the Vec.  Instead we can use Arc::from to create an Arc::<[T]>
directly which avoids the storage overheads.

Finally we need to update symbolic to get rid of the bad dmsort
dependency which doesn't work on rust 1.48.