feat(wasm): WASM support (#301)
This implements WASM support. It uses a slightly modified protocol where a sender can
supply an `addr_mode` to be explicit about the addressing mode. `"abs"` is the default
absolute addressing mode, another option now is `rel:index` with the index of a module
in which case the instruction address is considered to be module relative. This change
also now lets existing users benefit of this by providing the relative address within any
other debug module as alternative.