feat(difs): Provide a nonfound dif candidate when no files listed
When DownloadService::list_files returns no files for a source, none
is attempted to be fetched and you get no info on the DIF object
candidates.  This means you have no idea that no DIF candidates were
ever found if the source actually checks at list-file time (currently
only the sentry source, all other sources always generate a number of
locations to try).  In this case we now add a dummy NotFound entry to
the candidates for such a source.

This dummy entry will also be generated if a source is filtered out.
E.g. If the request is for ELF files and passing in a source that only
has PE/PDB DIF objects list-files will filter this out and not
generate any entries to fetch.  In this case you will also get a
NotFound entry for this source.  This is desireable because it
explains to users that certains sources returned no results.

It would be even better if the downloader could return a "filtered"
status instead of not found for this.  However getting more
information out of the downloader is out of scope for now and is for
future work.