Revert back to the old code
Seems like no-one (other than me) found the new code more readable.
It's a bit unfortunate that everything needs to be owned because it is
all spawned on a new threadpool, thus the new code doesn't become much
nicer.  Plust the future_metrics! macro didn't have an idiomatic
async-await substitute yet which doesn't help as it leaves everything
still in a half-way state.