ref(symsorter): Implement multi-threading (#342)
Uses `rayon` to move file handling and source bundle creation into a
thread pool. The main thread walks the given paths and continuously
produces paths into the thread pool. In the worker threads, the file is
opened, inspected, and potentially processed. The maximum number of
concurrent files at any given time therefore is `ncpu`.

Note that this is optimized for sorting directories with a high density
object files.