ref(downloader): Upgrade downloaders to tokio 1.0 (#335)
Introduces Tokio 1 in the downloader module, migrates to reqwest, and
updates to the latest version of rusoto for S3. Downloaders share the
same HTTP client and run in a single-threaded runtime. At this point,
the new runtime is only used by the downloaders.

Since reqwest does not allow to hook into the resolver, this uses the
patched version jan-auer/reqwest@62edd49 until reqwest adds the
required APIs.