feat: Update `url` to 2.2.0 (#353)
Updates the URL crate to 2.2.0. This allows us to pass the URL directly
to the web client without going through a second parse stage.

With version 2, the `url::percent_encoding` module is no longer exposed.
Instead, we have to use `Url::path_segments_mut` to construct a URL with
encoded path segments.