ref(utils): Replace tokio_retry with constant-time backoff (#357)
The tokio-retry crate has a dependency on an old version of the tokio
runtime. We do not need any tokio-specific functionality, only the
delay.

The exponential backoff was currently configured with a start delay of
10ms and limited to 3 retries in total. This does not result in
significant variance, so we can get rid of the exponential part
entirely. We can bring this back in a similar fashion with the backoff
crate however, which even has bindings for futures.