test: Port Minidump and AppleCrashReport tests to Rust (#351)
Ports some of the endpoint tests to Rust. Primarily, these tests are meant to
test the endpoints. For this reason, they do not use sources. Ultimately, we
need to investigate unit-testing the endpoint code without the full service in
the background.

However, the tests instantiate a full symbolicator application. With this, we
verify that middlewares and routing (via endpoint registration) are covered by
the tests.

One of the minidump tests used the Microsoft symbol server. To retain this test,
it was labeled as "integration" test, but remains in the endpoint module for
now. Once we determine a better strategy for integration and E2E testing, it can
move to its more permanent location.