fix(test): Enforce HTTP/1.1 to support proxying to MSDL (#380)
The Microsoft symbol server has started returning transfer-encoding: chunked
for its 404 responses. Our HitCounter proxy uses to return this header to the
caller, but at the same time responds with HTTP/1.0. This is invalid and
causes the HTTP client in Symbolicator to fail.

This problem does not occur in production code. The fix is to configure the WSGI
test server for HTTP/1.1.