fix(tests): Fix a race condition in integration tests (#388)
The mechanims used to select a free port for symbolicator to listen on
was racy and lead to flaky tests due to ports being in used.  This
insteads configures symbolicator to use port 0, ensuring that its
server gets a unique port number from the kernel without races.  The
port number is then extracted from the logs printed.

To do this the simplest way involves starting a thread, in fact this
is one of the modes that subprocess.Popen.communicate() uses
internally anyway so this isn't terrible.  Forever reading line by
line once started is perhaps not great but these are just tests after
all.