fix(cache): Update the mtime when it is 1h out of date
The code used to update the mtime wasn't actually updating the mtime
as nothing was ever written to the file, it was only opened.  This
pulls in a new library which seems to do a better job at this.

There is now an explicit unit test for this, before this was tried to
be covered by an extra iteration in an integration test but this was
occasionlly flaky and did not take into account the mtime needed to be
out of date for more than an hour.