fix(build): Update build order and build with latest Rust (#406)
Follow-up to #398 and #400, which aligns the build sequence with the
original script. The initial change lead to an empty binary.

Also switches from the rust/slim-stretch base image to
debian/stretch-slim and installs Rust manually. Since the official Rust
image is no longer updated, we have to install Rust ourselves to build
with the latest stable compiler. This is now required after #399.