feat(symcache): BCSymbolMap support in symbolicator (#403)
This adds support for fetching BCSymbolMaps from symbol servers
and uses them when building symcaches to resolve hidden symbols
in MachO objects.  For this it introduces a new auxiliary debug
information files (auxdifs) cache which is used to fetch and cache
BCSymbolMaps and the PLists required to find the DebugId of a
BCSymbolMap.  These new DIFs are only supported on the
unified symbol server layout.  These are also DIF files which
are not also object files, now the two terms are more distinct.

Symsorter is also updated to include the BCSymbolMap and
PLists.