fix(symcache): Better error reporting for bcsymbolmap failures (#409)
This logs a single sentry error for bcsymbolmap failures.  We lose the
log entry though.  But the error contains both the full error chain as
well as more context on what went wrong with it.

To make the functions more robust it changes the file_type arg to not
take a list, this way the error logging doesn't need to take care of
this and is also more type-strict.  It has some lifetime implications
in the downloader which gets cleaned up a little bit.