ref(api): Rename the PList filetype (#434)
This FileType is actually part of the public API as it can be used in
the configuration of a source to only use a source for certain
filetypes.  So we should have some meaningful name which won't be
wrong and confuse in the future without being too restrictive.
Calling it "PList" is just too generic, so that's a bad choice.

These changes are fine currently since none of this has been released
yet.