build(docker): Ignore other crates (#435)
We only want to build the main crate and ignore any others.  This
sadly means for now that we need to build without --locked but CI
already ensures that our dependencies won't change.

This also harcodes the Rust version, otherwise this is lost at the
mercy of docker's caching and never really upgraded.