ref(downloads): Always filter by filetype for sentry downloads
The sentry downloader was ignoring the filetypes and instead always
returned all types of files found.  The object actor (and now bitcode
actor) would then try and parse the files and mark them as malformed.

This results in many duplicate downloads for the wrong filetypes, instead
this always uses the filetypes to select downloads.