ref(sentry): Remove source tags from object sentry scope (#473)
The source tags are relevant to the download part, but having them
also part of the object scope means they end up in errros about the
processing like CFI errors and write errors.  Yet they have nothing to
do with them.  This means that when you search for the source tags you
currently get these unrelated errors.  So remove this.

The source tags for downloads are added in
<FetchFileDataRequest as CacheItemRequest>::compute() and are
unchanged here.