feat: Restore caching of download failures (#501)
This re-applies the changes in #484, with one difference: Download failures are
stored as `CacheStatus::Malformed` instead of `CacheStatus::Negative`.

This is meant to be a stopgap against repeatedly re-downloading files. A known 
caveat to the changes in this diff is that it is currently not possible in a 
cached failure to distinguish between download failures and malformed object 
files. A follow-up PR that attempts to address that drawback will be opened 
later.

Co-authored-by: Sebastian Zivota <loewenheim@users.noreply.github.com>