feat(ci): Run Sentry's Symbolicator integration tests (#500)
When we make changes to Symbolicator, we should test against the latest Sentry. Up until now, we would have to merge the change to Symbolicator's `master` branch and then open a PR on Sentry to see if anything would break.

This change ensures that every PR and `master` change is tested against the latest Sentry.