feat(cache): Stuff error messages into `Malformed` and `CacheSpecificError` cache files (#510)
Details about the offending error are now stored in the contents
of Malformed and CacheSpecificError cache files. This allows
diagnostics to be persisted beyond the lifespan of a particular
symbolicator task, and any follow-up cache hits may provide
more useful information as to why they received a non-positive
result.

This makes zero changes to the current behaviour of symbolicator
and simply adds in logic to begin storing and reading the error
messages without using them.