feat(cache): Use the `CacheSpecificError` cache status (#512)
This begins using the CacheSpecificError status in that
cache entries will be created using this status and they
will no longer be converted into Malformed cache entries.

As of this change, sentry will begin receiving more 
candidates with an error status. This does not introduce
any new types of statuses to object candidates.

CacheSpecificError entries do not propagate across 
caches in the same way that Negative and Malformed 
cache entries do. A CacheSpecificError in a download 
cache, for example will turn into a Negative cache entry 
in the CFI cache or the Sym cache.