ref: Avoid spawning separate download tasks (NATIVE-191) (#546)
Now that the main symbolicator runs on the IO pool, it is not necessary to spawn download tasks on that same pool and await those tasks. Just await those futures directly.
This also simplifies things around taking ownership of self and arguments.

This also introduces a `CancelOnDrop` wrapper which is wrapped around spawns that are otherwise never canceled when the outer future is being timed out.