ref: Expose the sentry HTTP headers in the context, not tags
Tags are indexed and we don't need to be able to search on these.
Plus adding a `sentry.project_id` in the tags is highly confusing wrt
request.scope.  Originally this stuff was exposed in the context as
part of the http headers, so lets at least do this partially again.