fix: Output correct function for client-side frames (NATIVE-150)
This uses the `function` property instead of `symbol`, which is ultimately what sentry uses.
It also avoids flagging modules as `unwind: missing` which were included in client-walked stacks, but which were *really* missing in the previous breakpad-based walk.

Additionally, it moves the error-reporting of invalid minidumps *after* the breakpad-based walk, to not log errors in case breakpad isnt able to process the minidump either.