Allow waiting for store operation to finish
Because store is not async and only sends a message the original
caller has no idea if it ever finishes.  This adds a oneshot channel
to send back a completion message.  It makes testing possible.

This also changes the tests to not return Result anymore, the errors
are still useless with them.  With .unwrap() the test failures at
least point to where the error came from.