ref(shared-cache): Do not block at process startup
Getting the first token may take a little while, this refactors
startup to do this is the background and not block during startup.
When the shared cache becomes available it will be used, it does not
matter much if there are a few cache uses before the shared cache is
available.