feat(shared-cache): Upload files on cache extensions as well (#648)
The local cache extends the lifetime of files by bumping the mtime at
most once an hour for files which keep being used.  There are a
surprisingly large number of cache entries which seem to never expire
and always stay in the local cache.

This is not ideal for the shared cache: its entries expire at some
point regardless of whether they are used or not.  And thus the really
common files can expire out, meaning it would be upon a new
symbolicator to repopulate these.

Instead this hooks up the shared cache to also attemp to upload every
time the local cache lifetime was extended.  This will normally only
result in requests but not in actual transfers of data because the
shared cache does not upload when a file already exists.  Only when a
file finally expires out of the shared cache will it be re-uploaded.