feat(shared-cache): Do additional exists check for refresh uploads (#650)
When refreshing uploads we make a lot of requests which we expect to
not be needed as the file probably already exists, GCS will not use
those uploads because we use ifGenerationMatch=0.  However as we use a
simple upload mechanism we are transferring the data anyway and that
takes time.

This instead adds an extra check if to see if the file already exists
in the shared cache.  This is racy, however it will already reduce the
duplicate transferred data a lot which should reduce the pressure in
the upload queue.