ref(shared-cache): Shorten the connect timeout
The metrics indicate that waiting for 3s is pointless, making a cutoff
at 500ms will make a few more requests timeout but that's a reasonable
tradeoff to not wait for 2.5s to get a timeout anyway on most slow
requests.

Some other time we could do some retries with circuit breaking maybe.
Fow now this is fine because the local cache exists as well.