ref: Serve the web endpoints from a dedicated pool
This goes back to our previous architecture that serves the web
endpoints in its own tokio runtime.  This makes sure that if the main
worker pool is starting to starve we still serve simple requests like
heartbeats.  This in turn is important since missing heartbeats means
k8s will kill the symbolicator pod which is a bit premature.