ci(danger): Do not require changelogs from dependabot
For plain dependency upgrades that require not changes we normally do
not need any changelog entries.  But not having them fails CI which
renders every dependabot PR a huge hassle.

Sadly we can not customise the dependabot PRs to add a #skip-changelog
marker (see e.g.
https://github.com/dependabot/dependabot-core/issues/2091), so lets
resort to changing the check instead.

There is a risk that someone pushes some changes to such a PR that
does require a changelog.  Hopefully we'll recognise this and do the
right thing anyway.