ci(danger): Allow skip-changelog markers in some comments (#714)
This allows the #skip-changelog markers in some comments as well.  Since
danger itself posts comments with the marker in we can not allow it from all
comments.  For now this simply accepts the marker if there's also a command
for dependabot in the same PR.

Additionally this tries to give danger permissions again to post a message in
the comments.