ref: Clean up empty cache directories
Thus far, the cleanup process never removed empty cache directories.
For one, the worklist based approach cleaning the parent before the children,
did not make this possible in the first place, but also because normal
operations would have errored out on failing FS operations, for example
when a directory or file could not be created in a non-existent directory.

This changes the cleanup process to be depth-first, cleaning up empty
directories as it goes. It also makes the writing side of things extra
defensive and retries FS operations in the unlikely case that a directory
is being removed while operating in it.