ref: Separate the symbolication Service from its HTTP Interface
Primarily separates out the symbolication service that does the actual symbolication from the symbolicator CLI and HTTP frontend.

This should make it easier to add alternative non-http-based frontends to symbolicator.
In the future, this can be refactored a bit further, moving the request/polling code to the http frontend as well, so *only* the actual symbolication remain in the symbolicator-service crate.

One drawback of the PR as it is, is the usage of a separate "test" feature to reexport the testing utilities to be usable in the http-focused crate.