ref: Move task measure, timeouts and SymbolicationError to frontend
Moves the timeouts and metrics for the symbolication tasks into the frontend.

This simplifies the `SymbolicationActor` further, also replacing the `SymbolicationError` with
`anyhow::Error`. The frontend still uses the enum to distinguish between timeouts and errors.

The metric names were adapted from the previous implementation, and the both the metrics and the timeouts measure something slightly different now.
They now measure (and timeout) the *complete* task, as opposed to parts of it.
I believe that especially with regard to the timeouts this is more correct, as previously the stackwalking of a minidump *and* the symbolication of it could both take up to an hour, so two hours in total.
Now each task is limited to an hour, no matter of the internal time distribution.
Similarly, this now measures the whole task instead of its parts separately.