ref: Update dependencies
Primarily updates the Sentry SDK, symbolic, and serde_yaml; as well as our forks of reqwest and cpp_demangle.

The released Sentry SDK version now supports profiling, so using a git ref for that is not necessary anymore.
The updated symbolic/cpp_demangle should slightly reduce the number of demangling failures that we see.
And the update of serde_yaml now explicitly rejects empty config files, which used to happen via the library previously.