ref: Load CFI from a ByteView not a CachePath
This removes the now unused `path` param to `CacheItemRequest::load` along with the `CachePath` type.
Similarly, it removed the `scope` param as well, as all the `CacheItemRequest` types have it on `self` anyways, and it was unused in most cases.
Also simplifies usage of sentry `to_scope`/`configure_scope`.

And lastly, fixes shared cache refreshes which are not happening anymore as the `expiration` was refreshed prior to refreshing the shared cache.