ref: Clean up Downloader retry logic (#926)
This adds a second iteration to the download-errors test, making sure
that round trips through the on-disk cache yield the same statuses.

This also tightens up the retry logic, as we should try things 3 times,
not an additional 3 times. It also adds a couple of notes for where the
retry logic does not match the expectation. Why are we retrying permission
errors that most likely will not change after retry, while we are *not*
retrying timeouts which we probably should, as an intermittent timeout
can be gone on the next try.