ref: Reduce cardinality of `futures::measure` metrics (#1012)
This removes a couple of tags that I don’t believe anyone has every looked at (certainly I haven’t), and which are also not used anywhere in DD.

The `source.download.X` metrics are tagged with the source type (or rather `source_metric_key`) already and are being displayed in a dashboard which should be the source of
truth for download related metrics.

I also question the usefulness of the `futures::measure` metrics in general. We use sentry performance monitoring for these as well through `tracing::instrument` which might be a much better way to look at these things in general.