feat: Add support for in-memory caching
Fixes #994 by adding the infrastructure and defaults for in-memory caching.

This implements weighing of cache items (based on a bunch of `size_of`s),
and per-item TTL based on the `ExpirationTime`.

Each cache defaults to ~100k in-memory size, which should be roughly ~1k items.
Except for `object_meta` which defaults to ~100M in-memory and is very hot,
and `cficaches` which defaults to ~400M and are expensive to parse.