Cache the result of Demangling
Demangling can be an expensive operation, which also happens for every frame and is not deduplicated for the same function names shared across threads.
Furthermore, we capture errors for demangling failures, to eventually be able to improve those. However we are being spammed with up to 8 of those errors a second, and they are highly duplicated too, making it difficult to actually find new cases we don’t handle yet.

So lets cache all the things, to avoid duplicated work and spamming S4S events.