Do not fail processing minidumps on invalid `code_file` (#1133)
Invalid base-names will be excluded from symbol server lookups, which will certainly not find anything useful. We still output the full (invalid) code file into the symbolicated event, as it might still be useful to customers.

As a drive-by change, this reverts the `capture_anyhow` change that captures real stack traces in S4S, as the stack traces are not that useful, and the grouping is worse.