Add a new `BundleIndex` for SourceMap processing (#1251)
Instead of querying the Sentry API for files that are needed during SourceMap processing,
the goal of the `BundleIndex` is that Sentry proactively feeds a `BundleIndex` to Symbolicator.

This index can then be consulted to determine which exact `ArtifactBundle` a certain file is contained in, or rather which `ArtifactBundle` contains the most recent version of a certain file.

Once the index is fully in place, Symbolicator will not need to query the Sentry API for any files anymore, as the `BundleIndex` fully covers the available files on Sentry. If a file is not in the index, Symbolicator can fall back to scraping in that case.