Refresh scraped SourceFiles in a regular interval
This adds the ability to cache-bust SourceFiles that are being fetched due to scraping is JS processing.
We do this by adding a hash to the scraping URL which acts as a cache key.

Additionally, this circumvents the shared cache for source files, as it will be ineffective in this case.