Introduce more metrics for JS File Lookup
In order to better measure the success of JS symbolication, this introduces more fine grained metrics that represent a funnel:

We capture metrics for source files and sourcemaps separately, with the following tags:
- found via debug-id
- found via release
- found via release-old
- found via scraping
- not found
- not needed (for sourcemaps)

This reuses existing metrics names as much as possible for now.

Additionally, this also captures more metrics for engineers that go into detail via which method the bundle was found.