Backfill missing DebugId from CFI Cache (#1305)
We have seen customer minidumps that do not have proper DebugIds in them.
However, there is a valid CodeId, which would allow the executable to be fetched from an external SSQP symbol server.
We transform this executable to a CfiCache, which in theory could have a breakpad `MODULE` entry. Though in practice it does not (yet), as we never output one.

This would allow backfilling the missing DebugID in the minidump with the one found in the executable that we use to generate the CfiCache.