Avoid using `moka` in the `SymbolProvider`
Profiling has shown that `moka` might be a bit overkill for the `SymbolProvider`.
The `SymbolProvider` is never used for long term caching, but only as a short term `HashMap`.
This removed `moka` from this code path, and also avoids a bunch of Clones and other work from the `process_minidump` flow.