Do not persist SourceMap contents in `FileInBundleCache`
According to metrics, the cache is quite effective in ST and on LPQ instances. Less so in production, most likely because the cache is too small.

To increase the effective size of the cache, we do not store the actual sourcemap contents in it anymore. These contents are only used for sourcemapcache generation, and are not usually required.