ref(proguard): Keep track of frame indices (#1411)
Symbolication can expand one raw JVM frame into multiple frames. When we return the symbolicated stacktraces, we want it to be possible to tell which symbolicated frames came from which raw frames. To this end, we add an index to frames that you need to pass in and that gets copied to the symbolicated frames.