ref(proguard): Properly size the in-memory cache (#1424)
By default, in-memory caches have a size of 100KiB. This is sufficient for file types where we only cache some bookkeeping structs in memory. In the case of Proguard, we cache ProguardMapper structs, and those can be quite hefty. In brief experiments I found a mapper to be slightly less than twice the size of the mapping file on disk, and those files can take up hundreds of megabytes.

Therefore:

* We estimate the memory footprint of a ProguardMapper as twice the size of the file on disk;
* We set the in-memory cache capacity for Proguard to 2GiB.